requirements.txt
fastapi
langchain
langchain-community
langchain-openai
openai
pydantic
python-dotenv
python-json-logger
supabase
tenacity
tiktoken
uvicorn
pytest
pytest-asyncio
pytest-mock